doc: updated credential schema format for v0.8-dev.2 (#114)
* doc: updated credential schema format for v0.8-dev.2

We need to use the following ID:
"$id": "vpr:verana:VPR_CHAIN_ID/cs/v1/js/VPR_CREDENTIAL_SCHEMA_ID",
mind the VPR_CHAIN_ID instead of the vna-devnet-1 or mainnet.

Now we are using the placeholder on the client side so the correct CHAIN_ID is resolved by the backend and replaced in situ.